Bzutil improvements, move execution parsing out to file
